delete cards refactoring action names
