+++ Block/Customer/Cards.php
@@ -41,6 +41,6 @@ class Cards extends Template
      */
     public function getDeleteLink($card)
     {
-        return $this->getUrl('omise/manage/delete', ['card_id' => $card['id']]);
+        return $this->getUrl('omise/cards/deleteaction', ['card_id' => $card['id']]);
     }
 }
