+++ view/frontend/layout/customer_account.xml
@@ -4,7 +4,7 @@
         <referenceBlock name="customer_account_navigation">
             <block class="Magento\Customer\Block\Account\SortLinkInterface" name="customer-account-navigation-my-credit-cards-link">
                 <arguments>
-                    <argument name="path" xsi:type="string">omise/manage/cards</argument>
+                    <argument name="path" xsi:type="string">omise/cards/listaction</argument>
                     <argument name="label" xsi:type="string" translate="true">Stored Credit/Debit Cards</argument>
                     <argument name="sortOrder" xsi:type="number">160</argument>
                 </arguments>
